requirements.txt
azure-ai-projects==1.0.0b10
azure-ai-inference[prompts]
azure-identity
azure-search-documents
pandas
python-dotenv
opentelemetry-api
azure-ai-projects
python-docx
fastapi
uvicorn[standard]
requests
streamlit